+++ Assets/Mochineko/YouTubeLiveStreamingClient/IAPIKeyProvider.cs
@@ -1,0 +1,11 @@
+#nullable enable
+
+namespace Mochineko.YouTubeLiveStreamingClient
+{
+    public interface IAPIKeyProvider
+    {
+        string APIKey { get; }
+
+        bool TryChangeKey();
+    }
+}